!!UNCHECKED!! New combat/character actions
I added a bunch of stuff to combat/character actions before I saw that the original combat file was corrected. Therefore lots of errors exist in this one still, but there's more stuff going on that we might want to import into the main branch. Feel free to have a look, experiment and upload to this branch.